requirements.txt
fastapi
uvicorn
python-dotenv
requests
openai
pillow
numpy
reportlab
moviepy

langchain
langchain-community
langchain-openai
faiss-cpu

beautifulsoup4
lxml
html5lib
